Update Services for Snow Leopard
Rename Services so that it is clearer what they do when they appear in
context menus:
    New MacVim Buffer With Selection
    New MacVim Buffer Here
The latter Service now works in Finder on Snow Leopard (it was broken
before).  The Service "Open Selected File" has been deprecated since it
does not seem very useful and it is rather confusing as to what it
really does.